Working --test benchmarks and public repo benchmark tests
