+++ Pipfile
@@ -8,6 +8,7 @@ jinja2 = "~=2.11.1"
 pytest = "*"
 pytest-benchmark = "*"
 semgrep = "*"
+pyyaml = "*"
 
 [packages]
