+++ tests/performance/test_public_repos.yaml
@@ -1,0 +1,8 @@
+go:
+- https://github.com/astaxie/beego
+python:
+- https://github.com/pallets/flask
+java:
+- https://github.com/mulesoft/mule
+javascript:
+- https://github.com/rtfpessoa/diff2html
